Rename to ni_measurement_plugin_sdk_generator
