+++ ni_measurement_plugin_sdk_generator/README.md
@@ -4,7 +4,7 @@
 
 ## Introduction
 
-MeasurementLink Code Generator for Python (`ni-measurement-plugin-generator`) is a
+MeasurementLink Code Generator for Python (`ni-measurement-plugin-sdk-generator`) is a
 tool for generating reusable measurement plug-ins using gRPC services.
 
 For installation and usage, see [Measurement Plugin Support for Python (`ni-measurement-plugin-sdk-service`)](https://pypi.org/project/ni-measurement-plugin-sdk-service/).
