+++ ni_measurement_plugin_sdk_generator/tests/conftest.py
@@ -1,4 +1,4 @@
-"""Shared fixtures for ni-measurement-plugin-generator tests."""
+"""Shared fixtures for ni-measurement-plugin-sdk-generator tests."""
 
 import pathlib
