+++ ni_measurement_plugin_sdk_generator/tests/test_assets/example_renders/measurement/measurement.py
